Fix vibrating windows
Some windows with just the right amount of scrollable content would trigger the
removal of the bottom margin, however the new layout pass would show that the
content is no longer scrollable so the bottom margin would be reenabled.
This would cause the window to continously vibrate up and down.

This fix verifies the available space before adding the margin, and doesn't
add it, if it would cause the content to scroll.