Fix crash when hiding keyboard before focus gain.
In some rare cases if the user accidentally dismisses a dialog before an
EditText widget gains focus, DialogUtils.hideKeyboard will instruct
InputMethodManager to hide the softkeyboard using the currently focussed
View, which is not yet available and therefore crashes.  This also
happens pretty much all the time during monkeyrunner tests, as the
series of events there is so quick and this race condition manifests
very easily.

This change attempts to use the dialog's View instance to get a window
token if the focus is not yet on any valid element, and if that is not
yet available for some reason, will not hide the softkeyboard (which
might be better than crashing in this case).